# Improve `Peep.Storage.Striped.get_all_metrics/1` performance