Hey @rkallos

The tests run fine. Kindly check if I haven't done anything stupid here. Thanks!

```
Operating System: Linux
CPU Information: AMD Ryzen Threadripper 1920X 12-Core Processor
Number of Available Cores: 24
Available memory: 31.22 GB
Elixir 1.16.0
Erlang 26.1.2
JIT enabled: true

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 14 s

Benchmarking default_dump ...
Benchmarking new_dump ...
Calculating statistics...
Formatting results...

Name                   ips        average  deviation         median         99th %
new_dump            909.39        1.10 ms    ±17.21%        1.04 ms        1.72 ms
default_dump         70.89       14.11 ms     ±4.02%       14.10 ms       15.49 ms

Comparison:
new_dump            909.39
default_dump         70.89 - 12.83x slower +13.01 ms
```

You can checkout the benchmark at: https://github.com/rkallos/peep/commit/8b02456bf8fb656c7a2d330ecd4b6c44d6c7b069