@aerosol 2025-05-07
(we could consider `Task.async_stream` to accumulate all the tids too, I guess?)