@aerosol 2025-05-07
FWIW this is to address something we observe a lot under heavy load:

```
15:36:39.386 [error] GenServer Plausible.PromEx.ETSCronFlusher terminating
** (stop) exited in: Task.await(%Task{mfa: {:erlang, :apply, 2}, owner: #PID<0.1849.0>, pid: #PID<0.729551.0>, ref: #Reference<0.0.236675.748871831.744030209.240037>}, 10000)
    ** (EXIT) time out
    (elixir 1.18.3) lib/task.ex:888: Task.await_receive/3
    (prom_ex 1.11.0) lib/prom_ex/ets_cron_flusher.ex:61: PromEx.ETSCronFlusher.handle_info/2
    (stdlib 6.2.2) gen_server.erl:2345: :gen_server.try_handle_info/3
    (stdlib 6.2.2) gen_server.erl:2433: :gen_server.handle_msg/6
    (stdlib 6.2.2) proc_lib.erl:329: :proc_lib.init_p_do_apply/3
Last message: :flush_ets
```

But trying out this branch now, it still appears sometimes.