@rkallos 2025-05-07
Wow, that's quite the improvement!

I haven't tried to replicate your results yet, and I'm curious to know what causes most of the slowdown. I'll try and find some time to look into this.

In the meantime, Credo complains about the level of nesting in the function heads for %Metrics.LastValue{} and %Metrics.Distribution{}. Please either add `# credo:disable-for-next-line Credo.Check.Refactor.Nesting` appropriately, or maybe add a `.credo.exs` file to your PR that changes the threshold for that check from 2 to 3.