@aerosol 2025-05-07
Oops, sorry, didn't realise credo needs to be satisfied, [e4e2d31](https://github.com/rkallos/peep/pull/39/commits/e4e2d317fe9d61d8097a1c7f2542e4459339feac)