@aerosol 2025-05-08
I wouldn't trust myself just yet... updated the benchmark with https://github.com/rkallos/peep/commit/bb6c1bdf77e36e698940e5385057ce430a89aa63 and the improvement is now negligible... Looking to understand what did I do wrong previously, but it looks like I'm misusing benchee?

```
##### With input large #####
Name                   ips        average  deviation         median         99th %
new_dump            1.03 K        0.97 ms    ±32.19%        0.91 ms        1.78 ms
default_dump        0.97 K        1.03 ms    ±32.19%        0.96 ms        1.80 ms

Comparison: 
new_dump            1.03 K
default_dump        0.97 K - 1.06x slower +0.0571 ms

##### With input medium #####
Name                   ips        average  deviation         median         99th %
new_dump            1.04 K        0.96 ms    ±16.14%        0.88 ms        1.52 ms
default_dump        0.98 K        1.02 ms    ±18.15%        0.91 ms        1.66 ms

Comparison: 
new_dump            1.04 K
default_dump        0.98 K - 1.06x slower +0.0606 ms

##### With input small #####
Name                   ips        average  deviation         median         99th %
default_dump        1.96 K      510.25 μs    ±20.20%      481.50 μs     1002.20 μs
new_dump            1.52 K      658.61 μs    ±16.48%      625.35 μs     1173.18 μs

Comparison: 
default_dump        1.96 K
new_dump            1.52 K - 1.29x slower +148.36 μs
```